requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
openai>=1.3.5
groq>=0.4.2
edge-tts>=6.1.9
streamlit>=1.37.0
requests>=2.31.0
websockets>=12.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
